requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
groq==0.15.0
structlog==24.4.0
alembic==1.13.3
httpx==0.27.0
celery==5.3.6
redis==5.0.8
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.12
